requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.6
djangorestframework==3.10.3
entrypoints==0.3
get==2019.4.13
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.3
opencv-python==4.1.1.26
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
public==2019.4.13
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
scikit-fmm==2019.1.30
scikit-learn==0.21.3
scikit-traveltime==0.0.7
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.4
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
